Safari-611.1.10/LayoutTests/http/tests/security/contentSecurityPolicy/source-list-parsing-08.html

<!DOCTYPE html>
<html>
<head>
<script src='resources/multiple-iframe-test.js'></script>
<script>
var dataURL = "data:application/javascript;base64," +
    btoa("var result = document.getElementById('result');" +
         "result.firstChild.nodeValue = result.attributes.getNamedItem('text').value;");
var tests = [
    ['yes', 'script-src data:', dataURL],
    ['no', 'script-src \'self\'', dataURL],
    ['no',  'script-src https://127.0.0.1:8000', dataURL],
];
</script>
</head>
<body onload="test()">
  <p>
    Test proper handling of data: URLs.
  </p>
